moved the reload check to listen to currentContext instead of filter. (#483)
Opening the fly out triggers a rerender and the reload routine starts firing while it doesnt need to.
The end result is the same here.
Removed listening to embedRef from the useEffect that triggers when isLoading changes. As it was causing unnesessery renders.
This useEffect should now only run on reload and initial load.