FilterPane header is now sticked to the top when scrolling. (#560)
FilterPane Header can now receive a component.
It will be aligned on the opposite side of the collapse-arrow.
Component will be hidden when collapsing filterpane.

Added a small filter header component to the Filter Pane story. Demoing how this works.
It just flips the filter pane placement back and forth from right to left.