fix to expandedContent by adding a max width. (#556)
Should now not expand the size of the table when opened.
Added a no wrap to expanded content label. so that it remains in one line if the row needs to wrap.
In util genrateColumnTemplte was changed for the selection and expanded content cells, making them min-content instead of max-content.
This because they were taking up space when opening up expanded content was opened, even if they were empty or didnt need more space.
causing the actually columns to be pushed further down.