BUG 11209 Profile - undefined text appears twice
if currentUsers given and family name is undefined, undefined is written out in title. Added a check for these and instead try to use fullName if we are not able to create a full name from given and family name. If fullname also does not exist, the title will be set to null and no name show. This makes sure we never get undefined in this component. Why currentUser ends up with no given or family name is not explored in this fix. (#624)