feat:HangingGarden component (#602)
New Component Hanging Garden. 
A component that give you the possibility to create a hanging gardin by supplying it instructions on how you want the garden to look. 
without having to worry about the complexity of rendering and the pixi.js library. 

With that said, this is a fairly complex component to implement. 
There are quite a few paramters that need to be filled out. 
The documentation for this is lacking for this release, but is being worked on.