reverted previous change where  Filter key to changed based on terms. As this caused the Filter to be recreated for each change/selection, making it loose its internal state. Most notably, this causes the filter to collapse each time a user makes a selection.
This changes was made to fix an issue where filters was not updating when user were applying Bookmarks.
I have made an alternative solution to this inside the Filter component itself.
It will now update based on defaultTerm. So if the filter selection get updated from outside the filter component. It will update its interal term state, reflecting these changes.