the reportInfo componeents css does not show when in use in apps. (#1285)
Tested this locally, but the problem was not very consistent, and kinda fixed itself without any changes.
In an attempt to fix this, Ive changed the parent className for reportInfo so its unique, in terms of the old PBI component. As we have had problems with this in the past.
Also moved the useElevataionClassname outside the render part of reportInfo.