Hotfix/powerbi checkcontextaccess (#1366)
* intial problem here was the checkContextAccess.
This was not reimplemented properly, and was run for all reports that implements the context.
but this service requres the report to also have RLS configuration.
Meaning all reports without RLS but who uses context, would fail the checkContextAccess check.

The fix for this turned out abit complicated, and also revealed some other minor issues. All should be adressed with this update.

checkAccessContext should only run if report "hasContext" and RLS.
To determine RLS, we need to fetch thee embedInfo.
This was previously not done until "hasContext" was set to true.
New routine always starts off getting the EmbedInfo.
On EmbedConfig change, we set HasRls accordingly.
When hasRls is set, we contiune to determined if we have contextAccess.
Where atlast the AccessToken request is fired when contextAccess is true.

This opened up more issues that has been resolved.
Unneceserry spinner on context switch.
When user changed context, checkContextAccess and requestAccessToken is run again. This caused a brief spinner on top of the report.
Added a silent mode to these request. So to not show spinner on subsequent request.

AccessTokenRequest forced report reload on context change.
When changing context, a new token was requested. This again forces the entire report to reload.
When hasContextAccess updated, it will only run if  hasContextAccess is true and no token is present.
The Report should no longer reload, and just filter itself in place as before.