fix(usePeopleDetails): useEffectAsync (#1474)
* fix(usePeopleDetails): useEffectAsync

Updating to use the async `useEffectAsync` from the Fusion package when updating the state with person data, as not doing this can result in errors like
```
react-dom.development.jss:68 Warning: Can't perform a React state update on an unmounted component. This is a no-op, but it indicates a memory leak in your application. To fix, cancel all subscriptions and asynchronous tasks in a useEffect cleanup function.
    at PersonPhoto
```
(here observed in the CPR app).

This change should mitigate the warning/error from React, as it no longer will attempt to perform state updates if the component becomes unmounted.

* fix(usePeopleDetails): Typo

Fixing typos which caused the build to fail: `signal.abort` -> `signal.aborted`
(did not catch this previously, as the editing of the file is being done directly in GitHub, rather than through an proper IDE).

* fix(usePeopleDetails): Signature useEffectAsync